+++ projects/ws2812/mcc_generated_files/pin_manager.c
@@ -1,0 +1,111 @@
+/**
+  Generated Pin Manager File
+
+  Company:
+    Microchip Technology Inc.
+
+  File Name:
+    pin_manager.c
+
+  Summary:
+    This is the Pin Manager file generated using PIC10 / PIC12 / PIC16 / PIC18 MCUs
+
+  Description:
+    This header file provides implementations for pin APIs for all pins selected in the GUI.
+    Generation Information :
+        Product Revision  :  PIC10 / PIC12 / PIC16 / PIC18 MCUs - 1.81.8
+        Device            :  PIC16F18313
+        Driver Version    :  2.11
+    The generated drivers are tested against the following:
+        Compiler          :  XC8 2.36 and above
+        MPLAB             :  MPLAB X 6.00
+
+    Copyright (c) 2013 - 2015 released Microchip Technology Inc.  All rights reserved.
+*/
+
+/*
+    (c) 2018 Microchip Technology Inc. and its subsidiaries. 
+    
+    Subject to your compliance with these terms, you may use Microchip software and any 
+    derivatives exclusively with Microchip products. It is your responsibility to comply with third party 
+    license terms applicable to your use of third party software (including open source software) that 
+    may accompany Microchip software.
+    
+    THIS SOFTWARE IS SUPPLIED BY MICROCHIP "AS IS". NO WARRANTIES, WHETHER 
+    EXPRESS, IMPLIED OR STATUTORY, APPLY TO THIS SOFTWARE, INCLUDING ANY 
+    IMPLIED WARRANTIES OF NON-INFRINGEMENT, MERCHANTABILITY, AND FITNESS 
+    FOR A PARTICULAR PURPOSE.
+    
+    IN NO EVENT WILL MICROCHIP BE LIABLE FOR ANY INDIRECT, SPECIAL, PUNITIVE, 
+    INCIDENTAL OR CONSEQUENTIAL LOSS, DAMAGE, COST OR EXPENSE OF ANY KIND 
+    WHATSOEVER RELATED TO THE SOFTWARE, HOWEVER CAUSED, EVEN IF MICROCHIP 
+    HAS BEEN ADVISED OF THE POSSIBILITY OR THE DAMAGES ARE FORESEEABLE. TO 
+    THE FULLEST EXTENT ALLOWED BY LAW, MICROCHIP'S TOTAL LIABILITY ON ALL 
+    CLAIMS IN ANY WAY RELATED TO THIS SOFTWARE WILL NOT EXCEED THE AMOUNT 
+    OF FEES, IF ANY, THAT YOU HAVE PAID DIRECTLY TO MICROCHIP FOR THIS 
+    SOFTWARE.
+*/
+
+#include "pin_manager.h"
+
+
+
+
+
+void PIN_MANAGER_Initialize(void)
+{
+    /**
+    LATx registers
+    */
+    LATA = 0x00;
+
+    /**
+    TRISx registers
+    */
+    TRISA = 0x15;
+
+    /**
+    ANSELx registers
+    */
+    ANSELA = 0x00;
+
+    /**
+    WPUx registers
+    */
+    WPUA = 0x14;
+
+    /**
+    ODx registers
+    */
+    ODCONA = 0x00;
+
+    /**
+    SLRCONx registers
+    */
+    SLRCONA = 0x37;
+
+    /**
+    INLVLx registers
+    */
+    INLVLA = 0x3F;
+
+
+
+
+
+   
+    
+	
+    SSP1CLKPPS = 0x01;   //RA1->MSSP1:SCK1;    
+    RA5PPS = 0x04;   //RA5->CLC1:CLC1OUT;    
+    RA1PPS = 0x18;   //RA1->MSSP1:SCK1;    
+    SSP1DATPPS = 0x00;   //RA0->MSSP1:SDI1;    
+}
+  
+void PIN_MANAGER_IOC(void)
+{   
+}
+
+/**
+ End of File
+*/