+++ docs/CHANGELOG.md
@@ -8,3 +8,5 @@ Changelog
 3. Third project battery tester added 15-06-2019
 
 4. Ultrasonic projects added 10-08-2020
+
+5. Ws2812 RGB LED project added 04-2023